+++ CHANGELOG.md
@@ -4,7 +4,7 @@ NeoForge-1.21.1-21.1.230
 
 #### ⭐ Improvements
 
-- Slime may now be crafted in the Composter [(\#230)](https://github.com/EnigmaticaModpacks/EnigmaticSkies/issues/230)
+- Slime may now be crafted in the Composter [(\#232)](https://github.com/EnigmaticaModpacks/EnigmaticSkies/issues/232)
 
 #### 🦟 Bugs Fixed